Fix for building on Linux
--HG--
extra : convert_revision : svn%3Acc11325d-82ef-9f48-b9d0-2d580b0792da/trunk%4094

+++ src/pylaunchy/python_export_qstring.cpp
@@ -105,7 +105,7 @@ struct QString_from_python_str_or_unicode
         QString tempString("");
         int i;
         for (i = 0; i < unichars; i++)
-            tempString.append(QChar(value[i]));
+            tempString.append(QChar((unsigned int)value[i]));
 #else
         // Python is using a 2-byte unsigned buffer of UCS-2 with
         // limited support for UTF-16
